requirements.txt
aioredis==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.6.2
cffi==1.16.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
daphne==3.0.2
Django==5.0.6
django-allauth==0.63.3
hiredis==2.3.2
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
jwt==1.3.1
msgpack==1.0.8
oauthlib==3.2.2
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
redis==5.0.5
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
stripe==9.9.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.1
urllib3==2.2.1
zope.interface==6.4.post2
